requirements.txt
Quart == 0.19.4
python-dotenv
aiohttp
asyncpg
requests